[REFACTOR] 디렉토리 분리 - 회원가입
