+++ src/main/java/com/balybus/galaxy/signUp/service/SignUpService.java
@@ -1,0 +1,21 @@
+package com.balybus.galaxy.signUp.service;
+
+import com.balybus.galaxy.global.domain.tblCenter.dto.CenterRequestDto;
+import com.balybus.galaxy.global.domain.tblCenter.dto.CenterResponseDto;
+import com.balybus.galaxy.global.domain.tblCenterManager.dto.CenterManagerRequestDto;
+import com.balybus.galaxy.global.domain.tblCenterManager.dto.CenterManagerResponseDto;
+import com.balybus.galaxy.global.utils.mail.dto.MailRequestDto;
+import com.balybus.galaxy.global.utils.mail.dto.MailResponseDto;
+import com.balybus.galaxy.login.classic.dto.request.HelperCertDTO;
+import com.balybus.galaxy.login.classic.dto.request.SignUpDTO;
+
+import java.util.List;
+
+public interface SignUpService {
+    MailResponseDto.AuthenticationMail authenticationMail(MailRequestDto.AuthenticationMail dto);
+    MailResponseDto.CheckAuthenticationCode checkAuthenticationCode(MailRequestDto.CheckAuthenticationCode dto);
+    List<HelperCertDTO> signUpHelper(SignUpDTO signUpRequest);
+    CenterResponseDto.GetCenterList getCenterList(CenterRequestDto.GetCenterList centerDto);
+    CenterResponseDto.RegisterCenter registerCenter(CenterRequestDto.RegisterCenter centerDto);
+    CenterManagerResponseDto.SignUpManager signUpManager(CenterManagerRequestDto.SignUpManager centerDto);
+}
